requirements.txt
selenium
pytest
python-dotenv
webdriver-manager
pytest-rerunfailures
allure-pytest